Refactor custom.js, UserSeeder, and AppServiceProvider
This commit refactors the custom.js file by removing the toast function. It also updates the UserSeeder to create an admin user with the 'superadmin' role and sets a default password. Additionally, the AppServiceProvider is modified to register policies and add a Gate before function to check for the 'superadmin' role.